+++ src/UI/IconButton/IconButton.tsx
@@ -1,15 +1,17 @@
 import { FC, ReactNode } from 'react';
+import cn from 'classnames';
 
-import styles from './IconButton.module.scss';
+import s from './IconButton.module.scss';
 
 interface Props {
   onClick: () => void;
   children: ReactNode;
+  hasBorder?: boolean;
 }
 
-export const IconButton: FC<Props> = ({ onClick, children }) => (
+export const IconButton: FC<Props> = ({ onClick, children, hasBorder }) => (
   <button
-    className={styles.iconButton}
+    className={cn(s.iconButton, { [s.border]: hasBorder })}
     type="button"
     onClick={onClick}
   >
