+++ src/components/Container/Container.tsx
@@ -7,8 +7,8 @@ interface ContainerProps {
   className?: string;
 }
 
-const Container: React.FC<ContainerProps> = ({ children, className }) => (
-  <section className={`section ${className || ''}`}>
+const Container: React.FC<ContainerProps> = ({ children, className = '' }) => (
+  <section className={`section ${className}`}>
     <div className={styles.container}>{children}</div>
   </section>
 );
