+++ src/styles/base/_base.scss
@@ -68,6 +68,11 @@ button {
   outline: none;
 }
 
+svg {
+  width: 16px;
+  height: 16px;
+}
+
 .visible {
   opacity: 1;
 }
